Rename janulator to swarmer.
