+++ swarmer/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'kotlin'
 apply plugin: 'application'
 apply plugin: 'org.junit.platform.gradle.plugin'
 
-mainClassName = 'com.gojuno.janulator.MainKt'
+mainClassName = 'com.gojuno.swarmer.MainKt'
 
 dependencies {
     compile libraries.cmdCommon
